# MyOffGridAI-Server-OpenAPI.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserDetailDto'
    put:
      tags:
      - user-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserDetailDto'
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/users/{id}/deactivate:
    put:
      tags:
      - user-controller
      operationId: deactivateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAuthResponse'
  /api/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAuthResponse'
  /api/auth/logout:
    post:
      tags:
      - auth-controller
      operationId: logout
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAuthResponse'
  /api/users:
    get:
      tags:
      - user-controller
      operationId: listUsers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListUserSummaryDto'
components:
  schemas:
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 100
          minLength: 1
        email:
          type: string
        role:
          type: string
          enum:
          - ROLE_OWNER
          - ROLE_ADMIN
          - ROLE_MEMBER
          - ROLE_VIEWER
          - ROLE_CHILD
    ApiResponseUserDetailDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserDetailDto'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    UserDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - ROLE_OWNER
          - ROLE_ADMIN
          - ROLE_MEMBER
          - ROLE_VIEWER
          - ROLE_CHILD
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
        email:
          type: string
        displayName:
          type: string
          maxLength: 100
          minLength: 1
        password:
          type: string
        role:
          type: string
          enum:
          - ROLE_OWNER
          - ROLE_ADMIN
          - ROLE_MEMBER
          - ROLE_VIEWER
          - ROLE_CHILD
      required:
      - displayName
      - password
      - username
    ApiResponseAuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/AuthResponse'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/UserSummaryDto'
    UserSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - ROLE_OWNER
          - ROLE_ADMIN
          - ROLE_MEMBER
          - ROLE_VIEWER
          - ROLE_CHILD
        isActive:
          type: boolean
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
      - refreshToken
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
      - password
      - username
    ApiResponseListUserSummaryDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserSummaryDto'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32