# src/main/java/com/myoffgridai/auth/dto/AuthResponse.java
package com.myoffgridai.auth.dto;

public record AuthResponse(
        String accessToken,
        String refreshToken,
        String tokenType,
        long expiresIn,
        UserSummaryDto user
) {
}