# src/main/java/com/myoffgridai/auth/dto/LoginRequest.java
package com.myoffgridai.auth.dto;

import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank(message = "Username is required")
        String username,

        @NotBlank(message = "Password is required")
        String password
) {
}