# src/main/java/com/myoffgridai/auth/dto/RefreshRequest.java
package com.myoffgridai.auth.dto;

import jakarta.validation.constraints.NotBlank;

public record RefreshRequest(
        @NotBlank(message = "Refresh token is required")
        String refreshToken
) {
}